Edit a bit in implementations

+++ content/numerical/Integrate.h
@@ -13,8 +13,7 @@
 
 template<class F>
 double quad(double a, double b, F f, const ll n = 1000) {
-	double h = (b - a) / 2 / n, v = f(a) + f(b);
-	rep(i,1,n*2)
-		v += f(a + i*h) * (i&1 ? 4 : 2);
+	double h = (b - a) / 2 / (double)n, v = f(a) + f(b);
+	rep(i,1,n*2) v += f(a + (double)i*h) * (i&1 ? 4 : 2);
 	return v * h / 3;
 }
